Complete prog-assign-2
Add function makeCacheMatrix that takes a numeric matrix as input and
returns a cacheMatrix object able to store both this matrix and its
inverse.
Add function cacheSolve that takes a cacheMatrix object as input and
returns the inverse of the corresponding matrix. It also stores that
inverse into the cacheMatrix object, in case it was not already cached.
Comment both functions.
Detail how to use them so as to avoid recomputing unnecessarily the
inverse of a matrix.